requirements.txt
azure-core==1.31.0
azure-data-tables==12.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
idna==3.10
isodate==0.7.2
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
propcache==0.2.0
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.3
shellingham==1.5.4
six==1.16.0
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.16.0
